remove a bunch of wip junk
