+++ session/manager_test.go
@@ -64,7 +64,7 @@ func TestItem_InvalidAt(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			mgr := &Manager[any]{}
+			mgr := &Manager{}
 			if tt.maxLifetime != nil {
 				mgr.opts.MaxLifetime = *tt.maxLifetime
 			}
